Lab 7 Answers:
1. The purpose of these auxiliary methods are to prevent code from inadvertently getting mixed
into other code. We want these methods to stay inside their intended and specific classes.
If the definitions from populateList, populateFifoList or ReverseLines.printReverse, than
the structure will be thrown off and not work properly.

2.These methods need arguments because each time through the recursive method, the
parameters may change. If the arguments meet a specified point, they will call the
return method, which returns the code (or loop you're in) back to the very method
who called it. The arguments change if the method sees that there is a null,
for example, or if the input has a nextLine. For each instance to be correctly
looked at (to see where to go next), the method keeps calling itself, but
with the different inputs from the user.

3. It is O(n)

4.populateList