What is the purpose of the various auxiliary methods populateList, populateFifoList, and ReverseLines.printReverse? These methods help programmers solve issues that have recurring steps. The populateList and populateFifoList methods repeatedly add items to a list. The printReverse method prints each and every item in a stack. These recursive methods wil continue by calling itself until the base case is reached rather than looping through each iteration.
Why do these methods need to have arguments, and how does the argument change from one recursive call to the next?
Argumentsare used in these particular methods to continually break down the problem into smaller/next steps until the base case is reached as the recursive call needs to be stopped when the current node is null for the populateList and populateFifo list methods. In addition, in this specific task, the argument gets updated every call to reflect the next node or input value.

What are the time and space complexity of each of the populateList populateFifoList methods, as well as ReverseLines.printReverse?
Complexities:

populateList is O(n) populateFifoList is O(n) ReverseLines.printReverse is O(n)

Which of these methods can be implemented using while loops?
All methods can be implemented and used using while loops.